+++ changelog.md
@@ -3,7 +3,8 @@ Naia - Project Ethea Phase 1 Shared Library - Changelog
 
 Version 20200808+dev:
 ---------------------
-
+* Lua and WML library:
+  * Added persistent WML variables API based on the AtS 0.10.x version.
 
 
 Version 20200808: